removed ResourceUtils
